+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr 0.3.0.99
 
+* Experimental `progress()` config which displays a progress bar, useful if
+  you're doing large uploads or downloads.
+
 * `PUT()`, `PATCH()` and `POST()` now use `encode` argument to determine how
   list inputs are encoded. Valid values are "multiple", "form" or "json".
   The `multipart` argument is now deprecated (#103).